+++ CHANGES.rst
@@ -8,6 +8,7 @@ Features:
  - keys backup: Implement setup screen (#2883)
  - keys backup: Display a warning on new sign out screen (#2885)
  - keys backup: recover screen #2887
+ - keys backup: Add a dedicated section to settings #251
 
 Improvements:
  - Support Split-screen mode (#1832)